requirements.txt
numpy~=1.21.5
Pillow~=9.0.0
torch~=1.11.0
torchvision~=0.12.0
albumentations~=1.1.0
tqdm~=4.62.3